Address some CodeQL warnings

+++ src/statement.h
@@ -86,7 +86,7 @@ public:
             callback.Reset(cb_, 1);
         }
         virtual ~Baton() {
-            for (unsigned int i = 0; i < parameters.size(); i++) {
+            for (size_t i = 0; i < parameters.size(); i++) {
                 Values::Field* field = parameters[i];
                 DELETE_FIELD(field);
             }
